AstrUtils officially starts
The configure and building infra structure (in the configure.ac and
Makefile.am files) is finally complete. I also corrected ImageCrop to
fit into this structure and it seems works very nice with in it. There
must be lots of problems I have not seen yet, but unfortunately I have
not noticed them yet. We will fix them as bugs ;-).

The revision history of the programs which where written before the
birth of AstrUtils (like ImageCrop) is not included in this revision
history. They will be modified and added, the revision history of new
programs and changes to the existing ones will however be included
automatically. I will keep the source files of the old programs with
their separate repository information for anyone who is interested.