AstrUtils changed to gnuastro and astr prefix to ast
AstrUtils was not a good project identifier, so after a discussion
with the GNU software evaluation team, it was changed to gnuastro. The
copyright notice in all the files and the manual had to be changed.

Also the prefix `astr` that was added to all the executables was
changed to `ast`.