ConvertType can now output to EPS and PDF
ConvertType can now save a complete EPS file and using GPL
GhostScript, it can also save the output as a PDF.

There are still some issues with color inputs and outputs which have
been explained in the ConvertType future updates section of the
manual. They will be dealt with later since most of the current work
with ConvertType will be on grayscale.