String literals instead of char *, system check
Before this, when ever a string literal was needed and the value could
be represented by a char * variable, the latter was used. To make sure
thereq might be no potential problem, all such cases were changed to a
"%s" and the name of the variable. It appears this could have been a
security problem.

Also all the calls to the `system` function where checked to see if it
really returned zero, if not the program will abort and report it.

In the manual two of the long sections in Introduction where moved to
the developing part.