CFITSIO types used instead of fixed width
Until now, the types of the arrays for CFITSIO functions were based on
the fixed width integer types. But CFITSIO operates based on "short",
"int" and "long" types!!!! The problem was actually seen when I was
trying to read a BITPIX=32 image. I had allocated 32bits for every
pixel, but CFITSIO would read and write 64 bits!

The manual was also updated.