Corrected flip bug in Convolve, gnuastro to Gnuastro
There was a bug in the kernel flipping part of Convolve: It should
only occur in the spatial domain and also there was a -1 missing in
the flip loop. They are both fixed.

After some discussions, we recognized that Gnuastro is better than
gnuastro in text. So all occurances fo `gnuastro' were changed to
`Gnuastro' in the manual.