libjpeg is now optional
Until now, if a user didn't have libjpeg, the configure script would
stop and there was no way forward. But since it is only used by one
part of one program (ConvertType), it was much more reasonable to have
as an optional requirement rather than a mandatory one. So in this
commit, all the places where libjpeg was needed is put within a macro
that is defined in config.h and set at configure time.