Continuing work on ImageWarp
The basic preparations for ImageWarp are now in place. All that
remains is to reverse warp the output pixels to the input pixels to
find the fraction of area that they cover on each pixel of the input
image to find what flux should be given to that output pixel.