Works if WCS structure could not be read
If WCSLIB fails to recognize the WCS structure of a FITS header, the
`readwcs' function in fitsarrayvv.c will now return a NULL pointer and
print a warning on stderr instead of aborting. In some cases the
structure might be irrelevant. So There is no need to stop the work if
it cannot be read. It is upto the programs which use this function to
check if the wcsprm structure was a NULL pointer or not.

Also continuing work on ImageWarp.