ImageWarp will correct WCS of input
Added corrections for the WCS structure of the input image so a point
in the input image now has the same world coordinates as a point in
the output image.

In the previous commit, the first pixel in the image had coordinates
(0.0f, 0.0f), now it has (1.0f, 1.0f).

Also removed the Gnuastro copyright information from the headers of
the output images since the Gnuastro copyright might have been
confused with the image copyright.