ImageWarp output in the same type as input
Until now the output of ImageWarp was in double type, but now it will
be in the same type as the input. The NaN pixels are also converted to
blank pixels for the integer data types.

To do this the `changetype' and `arraytofitsimg' in fitsarrayvv.c were
changed. The first will now round the floating point inputs to the
integer outputs and the second will also require a number of blank
pixels parameter.