MakeNoise initial version working
MakeNoise passed the initial run and was able to add noise to a mock
image.

All the occurances of `nul' or `NUL' in fitsarrayvv.[ch] were also
changed to `blank' or `BLANK'. This was done to be more conformal to
the FITS standard which defines a BLANK keyword for integer data
types.