Added test for MakeNoise, input info in headers
A test was added for MakeNoise.

The main array to fits functions in fitsarrayvv.c can now accept a
linked list of headers to add to the output. This was used to add the
input information to the headers of the outputs of ImageWarp and
MakeNoise so the results are more reproducable.