Fictional tutorials, but fixed in MakeProfiles
The two tutorials in chapter 2 (Tutorials) were re-written to add a
fictional story of some history of astronomy figures. This should make
the text easier to read for a new commer and also add some scientific
purpose behind the demonstrated application.

There was a bug in MakeProfiles where the last set of objects assigned
to a thread would not be written in the output and thus it would case
the `write' function to wait indefinitely, it is now corrected.

Also background value in MakeNoise was in units of flux before. In
practice since it relates to astronomical objects, it is more natural
(and understandable) if it is in units of magnitude. So we made this
change.