Titles and MathJax in html_node manual
The node by node HTML manual produced by Texinfo doesn't actually have
a title on each html. So the random reader (who comes in through a
search engine) will have no idea what this page is!!! So I added
titles in the start of these pages. I also added links to the other
formats of the manual in the bottom of the page and also to the main
Gnuastro webpage.

The manual extensively uses equations. So I included MathJax in the
pages that have equations so they are displayed nicely. I also added a
javascript.html page that lists the Javascript codes, sources and
licences that were used in Gnuastro's webpage to comply with tools
like LibreJS.