`(dir)' link in webpage corrected, minor corrections
In the top HTML pages (./manual/gnuastro.html and
./manual/html_node/index.html), the link `(dir)' would link to
non-existing places: `dir.html' and `../dir/index.html'. So I wrote a
simple SED script to set these two locations to the
./manual/index.html page.

In Circumference mode of MakeProfiles, when the profile was highly
elongated, the one pixel width circumference would not show near the
minor axis. To correct for this, when the truncation radius is reached
in this mode, instead of ignoring the pixel, it is filled in this mode
(for a circumference).