Spaces in HTML and manual corrected.
In the code that went through the HTML code to do the corrections in
forwebpage.sh, starting spaces would be removed (because readline just
starts from the first non-white character). This is now corrected with
the ` IFS='' ' line in before `read'.

In the manual, for Convolve and MakeNoise, the type of values some
options get was missed. They are now corrected.

The notice on the webpage (that the source code is not yet ready) was
also changed to a cartouche style to be more clear.