SubtractSky finds the sky on successful meshes
The first major part of SubtractSky is now added: It will find the sky
value on the mesh grid elements that can give a value (whose mode and
median are approximately equal). Interpolation and smoothing are still
left to be done.

To check the mode of the image, before continuing on to finish
SubtractSky, ImageStatistics will be added.