WCSLIB version, ASCII85 in EPS and minor corrections
Added support for wcslib_version in WSCLIB 5.5. Soon the other new
features of WCSLIB 5.x will also be implemented.

ASCII85 encoding was also applied for the EPS output of
ConvertType. This shrinks the storage space of each EPS image by
nearly 40 percent compared to Hexadecimal format. There is an option
for the user if they insist on Hexadecimal EPS output.

Made some small corrections to some functions, see the ChangeLog.