MakeProfiles also using RNG seed, several bug fixes
Now MakeProfiles will also use the random number generator's ability
to get the seed value from the environment to have completely
reproducible models.

In convolve.c, the output would show the spectrum of the final
(complex) convolved image, not only its real part. This was no problem
in mock images that had been tested until now. But for real images
(that have negative values), this resulted in only a positive
value. This bug is now fixed.