First verison of mesh interpolation complete
Interpolation can now be done on the mesh grid. There are two general
approaches, one is to interpolate over only one channel. The other is
to interpolate over the whole image (if multiple channels are
identified). Both of them seem to work nicely.

A tiny bug was also fixed in mode.c (modesymmetricity) where the topi
variable could be larger than the number of elements in the input
array. Now that is fixed.