Convolution on meshs (possibly independent of channels) complete
In order for the noise properties of each channel to not be mixed,
convolution would have be done independently on each mesh. A new set
of functions were added to mesh.c to do the convolution on each
channel independently. Ofcourse, there is also an option
(--fullconvolution) so if the user wants, the convolution will take
place over the full image irrespective of the channels.