External function for operating on each mesh
Until now, the function that would operate on the meshs was defined in
mesh.c. This made it very hard to do a variety of things. But now,
using a function pointer, the function that will be spinned off by
pthread_create is a function pointer in operateonmesh so any of the
programs can easily use their own functions. There is also a `void *'
pointer in meshparams that can point to the parameters structure of
the calling program so all the parameters of the program are
accessible.

The example in src/subtractsky/subtractsky.c can be a good tutorial.