NoiseChisel thresholding done, Convolution on mesh
NoiseChisel's thresholding step is now added.

In order to to add thresholding to NoiseChisel, convolution had to be
defined independently on each channel. So a convolution function was
added to mesh.c. Convolution on the meshs was also added to
SubtractSky and the spatial convolution in Convolve was also changed
to work based on channels.