NoiseChisel's detection is nearly complete
NoiseChisel's detection algorithm is nearly fully in place now. Both
the background regions and the detection regions are now processed and
the Signal to noise ratio is found for all their connected
components. The Signal to noise ratio limit is also found and smoothed
for all meshs. We just have to remove all the false detections in the
detected regions.

The Histogram and Cumulative frequency functions in statistics were
also corrected to work without any bugs. Until now, there was a
problem when a data element would fall on the last interval.