Corrections to the mesh.c garrays
The garrays in the mesh structure play a very important role: They
keep the values to be used on each mesh. However, the method I was
dealing with them before was very troublesome. Because in some cases,
it was necessary to use the full image and in other cases to use only
meshs on a channel.

All the functions in mesh.c were corrected to accurately account for
such issues. Now the caller doesn't have to know what the user asked
for (to use the meshes over a grid or in channels). All this is done
internally in mesh.c. The caller only has to use garray1 or garray2.