NoiseChisel's over-segmentation algorithm now in place
The over-segmentation algorithm is now in place and the segmentation
process is starting. So far only the clumps in the noise regions can
be found. Now, I have to add the functions to find the signal to noise
of the clumps, then find the S/N limit of the clumps and finally do
the same job on the detections to detect true clumps within them.