Fixed array for sky subtracted image in NoiseChisel
Until now, for each step of detection and segmentation, it was
necessary to make a temporary sky subtracted image. When the input
image is large, this will result in a large chunk of memory being
allocated and deallocated. It would also make the code a little
confusing.

Now a new member has been put in the noisechiselparams structure to
keep an array the size of the input image. This array will be used by
both the detection and segmentation processes and is allocated in ui.c
before any actual processing. Also thanks to this array (with a
different name) it is much more clear which one of the input image or
sky subtracted image we are dealing with.

Note that in particular for segmentation, using the sky subtracted
image to measure the total flux would make us add more noise (due to
the subtraction of the sky). So we use the actual input image.