Continuing work on NoiseChisel segmentation, bug in meshs fixed
Work is progressing in the segmentation of detections in NoiseChisel,
the detections are now over-segmented and their signal to noise is
calculated. I am now busy bringing in the removal of false clumps
function to find the true clumps.

While working on segmentation, I noticed that there is a problem in
the imgindextomeshid function. This function would get an image index
and output the ID of the mesh that pixel belongs to. The problem was
this: when the mesh size is not an exact multiple of the image size
(in any dimension), the last mesh on all dimensions will have a
different size. If it is larger, then before this correction, this
function would give one larger mesh index for the respective
dimension, which would result in reading values outside of the
garrays. Now there is a nice check.