False clumps can now be removed in NoiseChisel
The false clump removal process is now implemented in NoiseChisel. In
order to do it on a real image (with NaN values), I had to make some
modifications to the oversegmentation function. The list of top indexs
for each clump would not be accurate when there were NaN pixels in the
clump. So by setting some standards for that list and some checks in
the oversegmentation, it now stores the top non-NaN pixel of each
clump.