Corrected NoiseChisel options, added --meshbasedcheck
The fast writing of the NoiseChisel options was very buggy, so I went
through it again and made corrections.

While doing the corrections, I noticed an improvement: I could create
a function to write the mesh values to a FITS file instead of
separately writing checks everytime I wanted to do so. This idea
actually came to my mind when I wanted to add the --meshbasedcheck
option. With this functionality added, several checks would have to be
done multiple times in the course of one function which would make
things very hard to read for an outside reader. They were all also
repetative and could have caused bugs! These parts of the functions
are now much more cleaner and elegant than before.