MakeCatalog now reads the inputs
MakeCatalog's ui.c is complete for now (all the inputs are read). In
order to make things clear and elegant, I made a clear function to
check for data in other extensions of an image or in another file
depending on the user input.

Until now, there was a macro in checkset.h and also a function in
fitsarrayvv.c which done the job almost similarly. The mask checking
functions of all the programs were updated.