Polygon cropping in ImageCrop
ImageCrop can now crop a polygon from an input image. It is currently
written for both the WCS mode and image mode but has only been tested
for the image mode. Tomorrow, I will test it in WCS mode and write
tests for it too.

Also the keepmaxnearriver option was added to the configuration file
of NoiseChisel.