ImageCrop's outpolygon option now set and tested
I had forgot to confirm if the old `--inpolygon' option actually
worked! So I tried it out and saw that it is much better if it is
`--outpolygon'! Since by default only the region inner to the polygon
is shown, so the old method was mostly redundant. In the few cases
that the user would want to keep the outer parts, they would have to
go and change all their configuration files! Now this option is not in
configuration files any more.

A test was also written for this option.