Detection limit and depth expressed in catalog comments
The detection limit of the image for clumps and objects is now
reported in the output of MakeCatalog. In order to do this, I had to
save the S/N thresholds in NoiseChisel to header keywords in the
NoiseChisel outputs. MakeCatalog then reads those headers and makes
the information. The manual also contains a complete explanation.

I also had to update the old "readkeyword" function in fitsarrayvv.c
to "readkeywords". Now it can read as many keywords as the user wants.