Corrected numblank to anyblank
In the fitsarrayvv.c functions, I had mistakenly used a variable
called `numblank'. Some time in the past, this variable probably
actually did mean the number of blanks. But now it was only a boolean
value specifying if there are any blanks in the input or not. So I
renamed all wrong occurances of `numblank' (as relates to the input
image and FITS blank pixels) to `anyblank' which correctly represents
what this variable holds.

Note that NoiseChisel's segmentation does make use of a similarly
termed variable for the number of pixels left to grow. That is a
different usage and those variable names were left unchanged.