MakeCatalog: now correctly identifies river pixel information
To get the river information of each clump, I had mistakenly just
copied from the getclumpinfo function in NoiseChisel's
clumps.c. However, over there all the clumps were within one
detection. So the IDs of neighboring clumps (to a river pixel) were
unique. But in MakeCatalog, two clumps can be separated by one river
pixel and belong to separate objects with different object IDs.

Another consequence was to use the object ID from the neighbor of the
river pixel and not the object on which the river lies.

The total brightness of the clump is also stored internally. Only when
we want to write it to the catalog, do we subtract the average river
flux multiplied by the number of clump pixels.