Removed flux from MakeCatalog
I had made a mistake on the definition of flux. Flux is defined as the
area divided by the area that collects the light (the telescope
aperture), not the area you use to measure its area or the number of
pixels times the area of the pixels!!! So I removed flux from the
columns in MakeCatalog and corrected the explanation in the manual.