makekernel max radius, smaller output if necessary
The --makekernel option in Convlve, now gets an argument that is the
maximum radius to keep non-zero kernel values. The output kernel will
also be cropped to only be large enough to have no only-zero rows or
columns.

This was done because the high frequencies were set to zero during the
division and so the larger radii would become too flat. So the user
can specify how far out the final kernel should be.

Also a minor bug fix in NoiseChisel: The check to see if the number of
clumps or detections is more than the minimum specified by the user is
now done before anything else. In the case of zero good S/N values,
this would cause a problem in the next steps.