NoiseChisel Paper reproduction release
In order to exactly reproduce the paper, it is important to have a
version of Gnuastro and all its utitlies that were used to derive the
results of the NoiseChisel paper.

So a new version number `0.0' was defined in both configure.ac and all
the main.h files of all the existing utilities in order to not confuse
them with the first official release of Gnuastro (0.1) that will be
made together with this pre-release.