Same as commit 0e8bfe3a19ebedebdf3770861bfa548b173bfd68
It was necessary that the branch for the NoiseChisel paper
reproduction (version 0.0) would be separate from all the work that
was done on the programs after the last proof submission of the
paper.

After the paper was published, I completed the reproduction system and
in the process, some issues needed to be added or fixed in the
programs. They were all applied to that branch. But they are all also
necessary for the future versions of Gnuastro, so they were done
independently here.