MakeProfile's individual outputs have suffix
The individual outputs of MakeProfiles were simply named as, for
example, 0.fits, 1.fits and so on. But this would not work when
MakeProfiles is to be run in parallel and is also ambiguous. So now
the row numbers are suffixed by the output file name and put into the
output directory.

The output directory and file name fixing in MakeProfiles was also
bugy with problems I hadn't noticed before, so I fixed it.

In order to do this job, two functions were added in checkset.c:
`dirpart' and `notdirpart'.

A new subsection was also added to the manual explaining the two basic
methods to run processes simultaneously.