Added --nolog option to the common options
When the scripts are run in parallel (for example within a Makefile
called with the -j option) then the log files can cause problems when
you want to clean them up in the end. One option was to set the log
file name to be based on the output file name, but I have not seen
that commonly and it would require a lot of extra code. But now, when
the log file doesn't matter, the user can simply igore it being made.

Convolve was not reading the WCS information of the input. A line is
now added in the ui.c to do the job.