MakeCatalog: surface brightness limit and removed mag limit
Previously the surface brightness limit and magnitude limit were
defined in terms of a fixed projected area (arcsec^2) and the S/N of
the objects or clumps. But in the first case, the projected area
should not be used since noise is only dependent on the pixel and the
pixel is the unit of our data-collection. Its physical or projected
size are irrelevant. In the latter case, the area used to find the S/N
was not clear and thus this value was very unrealistic. Therefore I
tried to explain how to use number counts (from the outputs of
MakeCatalog) in order to estimate the limiting magnitude.