hstartwcs and hendwcs are available in ImageWarp too
Since ImageWarp is also a starting point in reading images with
non-standard WCS information, I also added this option to
ImageWarp. If it is needed in all applications, these options can be
added to the common arguments easily now. The main readfitswcs
function in lib/fitsarrayvv.[ch] was changed to allow these two values
for easy generalization if needed.