Header: Added a new --asis option, bug in MakeCatalog fixed
With this option, a user can easily read a full keyword string from
one FITS file and put it into another. Ofcourse, it can cause problems
if the string does not comply with the FITS standard, but I have tried
to explain (and warn and propose alternatives) thoroughly in the
manual.

I had mistakenly used "and" between the size checks of the different
images. So when only one side of the image was different, it wouldn't
give any errors. This is fixed now and there is an "or" between the
two axises.