Add bootstrap script to prepare the checked out distribution
Fixes task#13778.

We use the unmodified "bootstrap" script of gnulib to copy the needed
modules.

File "boostrap.conf" contains some configurations specific to our
project.  `bootstrap_post_import_hook' function caters for fetching
"m4/ax_pthread.m4" file from autoconf-archive repository.

The file README-hacking is needed by "bootstrap" to recognize this is a
checked-out distribution.  Actually, it should be a file present only in
the VCS source and not distributed in the release tarball, so it should
not be included in it.  For the time being the file is empty but we can
fill it later.  The name can also be changed, but the appropriate
variable should be also set in "bootstrap.conf" then.

AC_CONFIG_AUX_DIR([build-aux]) has been addedd to "configure.ac".

When there is the LT_INIT macro in "configure.ac", libtool is used by
the "bootstrap", and in that case the libgnu library has extension ".la"
instead of ".a".  Every "Makefile.am" has been updated accordingly.