First apparently complete version of ImageArithmetic
ImageArithmetic can now do addition, subtraction, multiplication and
division in any manner using the postfix or reverse polish
notation. Functions can easily be added and it is possible to
translate infix notation (commonly used) into postfix notation
later. There are certainly bugs which will be fixed as they are
found. A small test was also written and added and the manual was also
updated.