Arithmetic is new name for ImageArithmetic, sqrt added
Since Arithmetic can also work on numbers and in the future will also
be working on 3D images, the general name "Arithmetic" is more
suitable for it than "ImageArithmetic".

Also, the square root function was defined to simplify the job for the
users (instead of raising to the power of 0.5). Since it is a common
operation.