Arithmetic takes 20 inputs, manual intro updated
Previously Arithmetic would only take 10 input images, but now, it
takes 20 input images. This is a very ugly and temporary solution and
should change as explained in task #13887. I just didn't have the time
to implement that solution now.

The web address of the manual and the source code is now directly
written within the text of the first page. The main reason was that
some company (Samurai Media Limited) is sellling the manual on Amazon
and if anyone buys it from there, they have to know where to get the
most up to date version of the manual.