"manual" -> "book" for Gnuastro in manual
Since the manual is now being sold on Amazon, is very comprehensive
(far more than a standard manual) and is already nearly 200 pages, I
thought that maybe the word "book" might be more descriptive of it. So
I changed all the relevant "manual" words in the manual/book to
"book".

Some minor corrections were also made in the manual.