Some clarifications in Section 2.2 of the manual
Two minor issues were corrected/clarified in the second tutorial of
the manual. 1) The zeropoint was mistakenly explained (it does not
depend on the night of observation). 2) I put the explanation of how
to fill the cat.txt file to input into MakeProfiles. A new reader will
probably not understand trivially that they should just copy and paste
the output of `cat cat.txt` into the file when the editor opens.