CosmicCalculator now does the basic calculations
The most basic calculations and set of cosmological variables are now
done with CosmicCalculator. There is certainly place for many more
calculations in the future, but for now, this does seem to be a good
start. A very simple initial test is also written for it to just check
if it runs, later more sophisticated tests (which actually check the
values) have to be written.