MakeCatalog: moved RA-Dec calculations in position function
The RA-Dec calculations were done in the top main function. This was
not the right place, because the top function should not do such low
level steps. So I moved them to the position() function in
columns. They are now calculated only when needed.

I also started working on the second order moments for elliptical
measurements, most the required information is now stored in the first
and second pass, but not yet reported.