MakeCatalog: Brightness related columns also corrected
In the previous commit, the positional column corrections were
completely moved out of the first and second pass functions and the
corrections were called by the functions making the columns. In this
commit, this system is now implemented for the brightness
columns.

Things were also made much more clean and lots of comments were added
to make things more clear for someone inspecting the code.