MakeCatalog: Now produces elliptical parameters
MakeCatalog now calculates the semi-major axis length, the semi-minor
axis length and position angle of the profile both using in geometric
(ignoring pixel values) and pixel-value-weighted measurements. Nearly
everything is also calculated for the clumps-in-object elliptical
parameters too, but they are not yet included as options.

The addition of 1 to the position values (which is necessary due to
the FITS definition) is now moved within the first and second pass
functions, because the second order moments are no longer linear, we
couldn't make the corrections later.