MakeCatalog: internal shift in second order calculations
Floating point rounding errors can be catastrophic when comparing sums
of squares with simple sums (which is done in the second order
moments). This is especially the case for MakeCatalog, where an
object's X or Y in the image can be very large, producing huge sums of
squares.

The first time an object or clump is encountered, its X and Y is used
as a reference and all second order moments are calculated after
shifting to that reference point. Later, when correcting the second
order moments, the same shift is subtracted from the mean value to
make the calculations accurate. In this way, the mean and variance
will be roughly similar and so we can avoid floating point errors.