CosmicCalculator manual: Revised explanations, 3D added
The preliminary explanations on distances in a curved space (in 2D)
where not very complete, they were just a first writing. So going
through them, I edited/corrected the text and also added the
explanations for 3D distances. The distances were the hardest part to
explain for physical insight. Now that it is done, the rest of the
derivations can easily be done and will be added in due time.