MakeCatalog: Sky and its STD are now independent of threshold
The Sky and its Standard deviation are now calculated independent of
the threshold which is how it should be. Previously, I had mistakenly
only calculated the average Sky and its standard deviation for pixels
above the threshold. To do this, a new column in the object
information table was necessary: the OALLAREA column.