MakeCatalog: --noriverbrightness added for clumps
This option allows the calculation of the brightness of a clump
without subtracting the average river brightness, which is useful when
the clump brightness will be subtracted later, as explained in the
manual.

Also, in preparearrays, when we are setting the columns of the output
in the clumps or object arrays, the value to be put is the same value
that is being `switch'ed. So having different values after the equal
sign was redundant and could cause bugs. So I also corrected this.