Manual: Some additions to "Science and its tools"
As the manual is growing larger and is trying to thoroughly explain
all the concepts, I thought it is necessary to highlight its role in
this important section of the manual. The fact that it explains the
concepts thoroughly while also providing the code (as part of the
software) reminded me of Numerical Recipes, which I thought would be a
nice comparison for most readers who might be familiar with that book.