lib/checkset.c: Removed the checkfile from automaticoutput
The main work in this commit was to remove checkfile from the
automaticoutput function. automaticoutput was mainly designed for a
filename creation, having to check if the file exists causes some
complications which don't exist any more. It also caused redundant
checks when several automaticoutputs were called. So all the calls to
automaticoutput were checked an checkfile was called before
automaticoutput when ever needed.

The change in NoiseChisel (explained below) was actually the main
motivation for the above change in automaticoutput. These should have
been done in other commits, it will not be repeated again ;-)!

 - [NoiseChisel] Removed the --checkdetectionsn and --checkclumpsn
   which were no longer used (a single S/N is used for the full
   image). Also the --segsnhistnbins option was renamed to
   --clumpsnhistnbins.

 - [Documentation] Added the verion number in the file names for
   WCSLIB installation instructions.