Arithmetic: Accepts all input HDUs with single --hdu option
Finished task #13887. Now all the callings to the --hdu option will be
recorded for later use. Using a Preprocesser macro, it is now possible
for any program to not use the default commonargs.h, --hdu option. It
might later be used for other common options too. Now there is no
limit on the number of images that Arithmetic can work on.

Some other minor issues were also corrected here:

 - doublesumnum (in statistics.h) can now work reasonably with any NaN
   elements: if there are no elements or if all elements are NaN. This
   was mainly done for Arithmetic: for the parts that might all be NaN
   pixels.

 - mediandoubleinplace (in statistics.h) similar to doublesumnum
   above.

 - All the --hdu1..N were removed in Arithmetic.