Book: Developing sections after new gnuastro-devel mailing list
A new mailing list (gnuastro-devel) was defined on Savannah and all
the updates to the trackers on the project webpage were instructed to
forward their activities to it. This will allow all developers to be
up to date on activity that is being done, even if they are not
directly subscribed to a specific activity. It will also be good for
people that are not necessarily developers to follow the development.

The text in the top mailining list webpages were also updated (not
version controlled) to link to all the others and have a better
explanation. While doing that, the "Report a bug", "Suggest new
feautre" and "Gnuastro project webpage" sections were also updated
along with the top Gnuastro webpage HTML file.