Installation: Automatically build the AUTHORS file
A script (genauthors) was added to automatically build the AUTHORS
file from the version controlled source. This script should ideally be
run once before autoreconf and once before generating a distribution
tar file (to update AUTHORS). In any case, it must be a separate
script.

For now the AUTHORS file is still in the respository and updated to
what this script makes. But after the bootstrapping is done (to
prepare the version controlled source for ./configure), the AUTHORS
file will be removed, since it is automatically generated.