Installation: Added Gnulib's bootstrap script
Gnulib's bootstrap script was included along with a modified version
of bootstrap.conf. With these, once you have the source code and
Gnulib and autoconf-archive cloned, you just have to run ./bootstrap
and everything will be ready to run ./configure afterwards.

Mosè first came up with this idea and created an initial version of
this process. Learning from his work and studying the ./bootstrap
script my self, I added some new steps and added a lot of comments so
someone who is not familiar can easily grasp the idea by reading the
comments in bootstrap.conf.