Installation: Bootstrap process complete
Finished task #13778. Bootstrapping is the process of preparing the
cloned version controlled source for standard configuration and
building. Based on the great work done by Mosè on this task, I redone
the process (with many comments added) and Gnuastro can now be easily
compiled after clonning with the simple ./bootstrap command. Thanks a
lot Mosè for the great work.

While I was testing the final steps, I noticed that CFITSIO has also
been updated and now has "make utils" which will compile fpack and
funpack along with some other packages, so I also updated that part of
the catalog.