Bootstrapping: Updated README-hacking and genauthors, updated book
Some minor corrections were made in these two files to make them more
clear. Also a separate "Bootstrapping" section was added to the book,
explaining everything as nicely as I could right now, certainly it
will become much better in the future. But it needed a good
explanation, since this is not something done too often.