Book: Added an undoing bootstrap cartouche
A short cartouche was added to the Bootstrap section of the manual
explaining how to remove all the non version controlled files. This
can be a good guide for people that are not already familiar with the
git clean command.

Also, in the "Production workflow", it is now explained that all
commit authors will be included in the AUTHORS file.